requirements.txt
# alembic==1.16.2
# annotated-types==0.7.0
# anyio==4.9.0
# bcrypt==4.3.0
# cffi==1.17.1
# click==8.2.1
# colorama==0.4.6
# contourpy==1.3.2
# cryptography==45.0.4
# cycler==0.12.1
# dnspython==2.7.0
# ecdsa==0.19.1
# email_validator==2.2.0
# exceptiongroup==1.3.0
# fastapi==0.115.12
# fonttools==4.58.4
# greenlet==3.2.3
# h11==0.16.0
# httptools==0.6.4
# idna==3.10
# iniconfig==2.1.0
# Jinja2==3.1.6
# kiwisolver==1.4.8
# loguru==0.7.3
# Mako==1.3.10
# MarkupSafe==3.0.2
# matplotlib==3.10.3
# numpy==2.2.6
# packaging==25.0
# pandas==2.3.0
# passlib==1.7.4
# pdfkit==1.0.0
# pillow==11.2.1
# pluggy==1.6.0
# psycopg2==2.9.10
# pyasn1==0.6.1
# pycparser==2.22
# pydantic==2.11.7
# pydantic-settings==2.9.1
# pydantic_core==2.33.2
# Pygments==2.19.1
# pyparsing==3.2.3
# pytest==8.4.0
# python-dateutil==2.9.0.post0
# python-dotenv==1.1.0
# python-jose==3.5.0
# python-multipart==0.0.20
# pytz==2025.2
# PyYAML==6.0.2
# rsa==4.9.1
# seaborn==0.13.2
# six==1.17.0
# sniffio==1.3.1
# SQLAlchemy==2.0.41
# starlette==0.46.2
# tomli==2.2.1
# typing-inspection==0.4.1
# typing_extensions==4.14.0
# tzdata==2025.2
# uuid==1.30
# uvicorn==0.34.3
# watchfiles==1.0.5
# websockets==15.0.1
# win32_setctime==1.2.0


alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
httptools==0.6.4
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jwt==1.3.1
kombu==5.5.4
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pdfkit==1.0.0
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
vine==5.1.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
win32_setctime==1.2.0